+++ Snakefile
@@ -23,7 +23,7 @@ rule match_relevant_kmers_shap_values:
     output: 
         expand("data/matches/{feature_method}/{kmer}mers.csv", feature_method="shap_values", kmer=config["KMER"])
     shell: 
-        "python3 match_relevant_kmers.py {params.feature_method} {params.relevant_kmers_to_match}"
+        "python3 src/match_relevant_kmers.py {params.feature_method} {params.relevant_kmers_to_match}"
 
 # saliency maps
 rule match_relevant_kmers_saliency_map:
@@ -37,7 +37,7 @@ rule match_relevant_kmers_saliency_map:
     output: 
         expand("data/matches/{feature_method}/{kmer}mers.csv", feature_method="saliency_map", kmer=config["KMER"])
     shell: 
-        "python3 match_relevant_kmers.py {params.feature_method} {params.relevant_kmers_to_match}"
+        "python3 src/match_relevant_kmers.py {params.feature_method} {params.relevant_kmers_to_match}"
 
 ## 11. svm experiment
 # shap values
@@ -50,7 +50,7 @@ rule svm_shap_values:
     output: 
         expand("data/svm/{feature_method}/results_svm_{kmer}mer.csv", feature_method="shap_values", kmer=config["KMER"])
     shell: 
-        "python3 svm_experiment.py {params.feature_method}"
+        "python3 src/svm_experiment.py {params.feature_method}"
 
 # saliency map
 rule svm_saliency_map:
@@ -61,7 +61,7 @@ rule svm_saliency_map:
     output: 
         expand("data/svm/{feature_method}/results_svm_{kmer}mer.csv", feature_method="saliency_map", kmer=config["KMER"])
     shell: 
-        "python3 svm_experiment.py {params.feature_method}"
+        "python3 src/svm_experiment.py {params.feature_method}"
 
 ## 10. feature importance methods
 # shap values
@@ -73,7 +73,7 @@ rule shap:
         expand("data/shap_values/{clade}/relevant_kmers.csv", clade=config["CLADES"]),
         expand("data/shap_values/{clade}/shap_values.npy", clade=config["CLADES"]),
     script:
-        "shap_values.py"
+        "src/shap_values.py"
 
 # saliency map
 rule saliency_map:
@@ -84,7 +84,7 @@ rule saliency_map:
         expand("data/saliency_map/{clade}/representative_FCGR.npy", clade=config["CLADES"]),
         expand("data/saliency_map/{clade}/saliency_map.npy", clade=config["CLADES"]),
     script:
-        "saliency_map.py"
+        "src/saliency_map.py"
 
 ## 9. generate plots
 rule plots:
@@ -96,7 +96,7 @@ rule plots:
         expand("data/plots/accuracy_{kmer}mer.pdf", kmer=config["KMER"]),
         expand("data/plots/loss_{kmer}mer.pdf", kmer=config["KMER"]),
     script: 
-        "plots.py"
+        "src/plots.py"
 
 ## 8. clustering metrics
 rule clustering_metrics:
@@ -106,7 +106,7 @@ rule clustering_metrics:
     output: 
         "data/test/clustering_metrics.csv"
     script: 
-        "clustering_metrics.py"
+        "src/clustering_metrics.py"
 
 ## 7. classification metrics
 rule classification_metrics:
@@ -118,7 +118,7 @@ rule classification_metrics:
         "data/test/accuracy.txt",
         "data/test/curve_pr.pdf"
     script: 
-        "classification_metrics.py"
+        "src/classification_metrics.py"
 
 ## 6. test model
 rule test_model:
@@ -128,7 +128,7 @@ rule test_model:
     output: 
         "data/test/embeddings.npy",
         "data/test/predictions.csv"
-    script: "test.py"
+    script: "src/test.py"
 
 ## 5.  train model
 rule train_model:
@@ -139,7 +139,7 @@ rule train_model:
         "data/train/training_log.csv",
         "data/train/preprocessing.json"
     script: 
-        "train.py"
+        "src/train.py"
 
 ## 4. train, val, test sets
 rule split_data:
@@ -149,7 +149,7 @@ rule split_data:
         "data/train/datasets.json",
         "data/train/summary_labels.csv"
     script: 
-        "split_data.py"    
+        "src/split_data.py"    
 
 ## 3. Generate FCGR
 rule generate_fcgr:
@@ -158,7 +158,7 @@ rule generate_fcgr:
     output:
         expand("data/fcgr-{kmer}-mer/generated_fcgr.txt", kmer=config["KMER"])
     script:
-        "fasta2fcgr.py"
+        "src/fasta2fcgr.py"
 
 ## 2. Extract undersampled sequences in individual fasta files 
 rule extract_sequences:
@@ -167,7 +167,7 @@ rule extract_sequences:
     output:
         expand("data/{specie}/extracted_sequences.txt", specie=config["SPECIE"])
     script: 
-        "extract_sequences.py"
+        "src/extract_sequences.py"
 
 
 ## 1. Undersample sequences from metadata
@@ -179,4 +179,4 @@ rule undersample_sequences:
         "data/train/selected_by_clade.csv",
         "data/train/available_by_clade.csv",
     script: 
-        "undersample_sequences.py"
+        "src/undersample_sequences.py"
