+++ src/parameters.py
@@ -1,0 +1,4 @@
+import yaml
+
+with open("../parameters.yaml") as fp: 
+    PARAMETERS = yaml.load(fp, Loader=yaml.FullLoader)
