+++ src/train.py
@@ -4,8 +4,8 @@ import numpy as np
 import tensorflow as tf
 from sklearn.utils import class_weight
 
-from src.model_loader import ModelLoader
-from src.data_generator import DataGenerator  
+from loaders.model_loader import ModelLoader
+from loaders.data_generator import DataGenerator  
 from preprocessing import Pipeline
 from parameters import PARAMETERS
